+++ molgenis-ontology-core/src/main/java/org/molgenis/ontology/ic/TermFrequencyService.java
@@ -2,7 +2,7 @@ package org.molgenis.ontology.ic;
 
 public interface TermFrequencyService
 {
-	abstract Double getTermFrequency(String term);
+	abstract Float getTermFrequency(String term);
 
 	abstract Integer getTermOccurrence(String term);
